moved default logger codes to separated file
